Develop (#48)
* dockerize in php artisan serve

* dockerized with nginx php-fpm

* Update README.md

* Complete unit test for user route

* make 2nd db container, db run but app dont goes to test when running test

* Complete test unit for department

* complete unit test for employee

* complete test

* Add ci.yml

* upgrade CI PHP version to 8.3

* add migration env in ci.yml

* add wait for mysql

* another ci test

* fixing db for test phase

* fixing render view test failure

* fixing render view test failure

* fixing render view test failure

* fixing vite manifest missing

* remove --debug

* use artisan test

* Intentionally break CI test

* test signed commit

* test with good test

* fix ci.yml to do test when PR on devops/cicd

* make develop tested by CI

* Test/testing ci (#14)

* testing CI

* now test is good and also testing verified merge

* fix ci.yml to test on master not main (#16)

* add pdo_pgsql in dockerfile (#18)

* fix pgsql in dockerfile (#19)

* Fix/nginx (#20)

* fix pgsql in dockerfile

* fix nginx, dockerfile and entrypoint

* fix dockerfile (#21)

* fixing entrypoint (#22)

* add install gettext in dockerfile (#23)

* fix 502nginx (#25)

* fix permission issue (#26)

* chown cant be used because non root (#27)

* give permission on build not runtime (#28)

* fix permission (#29)

* install npm (#30)

* fix not correct view displayed (#32)

* fix viewpermission (#33)

* Update Dockerfile (#34)

* Update Dockerfile

* Update Dockerfile

* Update Dockerfile

* Update Dockerfile

* update Dockerfile

* update Dockerfile

* focing https

* update bootstrap/app.php

* revert change

* make forcehttps as middleware

* update ForceHttps

* update ForceHttps.php

* add schedule

* fix schedule import

* delete schedule

* fix UserController/destroy()

* add session error on user.index

* fix redirect

* fix local environtment container (#35)

* fixing dockerfile.demo (#37)

* Develop (#38)

* fix local environtment container (#35)

* fixing dockerfile.demo (#37)

* add demo version (#39)

* fix navigation on mobile (#41)

* fixing responsive for mobile view (#44)

* Fix/installation (#47)

* fix gitignore

* fix docker local, make it ready to run

* fix env

* Revise README for Dockerized setup and formatting

Updated README to enhance Docker setup instructions and improve formatting.

+++ app/Http/Controllers/UserController.php
@@ -76,6 +76,10 @@ class UserController extends Controller
     public function destroy(string $id)
     {
         $user = User::findOrFail($id);
+        if($user->email === "admin@example.com"){
+            return redirect()->route('user.index')->with('error', 'cannot delete the demo admin account');
+        }
+
         $user->delete();
 
         return redirect()->route('user.index')->with('success', 'User deleted successfully.');
